requirements.txt
pytorch_lightning
nltk
transformers
scikit-learn
numpy
more-itertools
jiwer
openai
matplotlib
networkx
pydot
sacrebleu
